fix: use correct stop ID for Back Bay tracks 5+7
Since Commuter Rail stop IDs are now based on track distance markers,
the stop ID prefix for tracks 5 and 7 at Back Bay is different from the
others, because these tracks are on a physically separate branch. This
fixes boarding status not being available for Worcester Line trains
(which use these tracks) because we were publishing it under the wrong
stop IDs.